refactor: update Neovim setup in install script
This commit modifies the install script to check if the Neovim
configuration directory is a symbolic link and if it points to the
correct location. If not, it clones the dotvim repository into a new
location and creates a symbolic link to it. This change ensures that
the Neovim configuration is always linked to the correct location.

Prior to this change, an early setup step was creating `~/.config/nvim`
(when we confirmed `nvim --version` after installation), so we weren't
actually getting the neovim config setup properly.