Crate local targets: Take 3!
In this version, we create a new build utils library that also includes
a binary to run which leverages `cargo metadata` to iterate all of the
packages within the workspace finding all of the binaries generated for
each and symlinking those binaries into the crate specific target directories.

This is nice because:

- we run it when we want to
- it's just a simple / normal binary itself -- easier to debug than
  build.rs
- it has no impact on build speed at all (since it's ran on demand)

The downsides are basically that we need to remember to run the command
after adding a new `src/bin` file (which is likely to be fairly rarely).