fix: Ensure multiple windows are started
The `ensure_window` function was checking the current state to decide if
it should call `new-session` or `new-window`, but we were never
updating the current state. This mean that we basically thought all
windows were new sessions (even though they weren't) and the behavior in
`tmux new-session` is that if the session specified already exists
nothing happens.