Use `diff.summary()` (it's better supported by neovim highlighting)
`diff.stat(72)` is good too (I actually like it better), but the
built-in neovim highlighting doesn't support it well. I looked into
adding it, but it's not really worth it. If I do want to move forward
there, it would be better to make an actual tree-sitter grammar for it
and support all of the things that might be inside the template.