feat(gitcommit prompt): support extracting diffs from commit message buffers
The `get_diff` function will now attempt to extract diff content
directly from commit message buffer if neovim was launched explicitly to
edit a `gitcommit`, `jj`, or `jjdescription` file (e.g. `nvim
.jj/repo/editor-asdfa.jjdescription`). If we don't find a diff in the
content, it just falls back to the same shell out logic as before.

Previously, we always assumed that `git diff` or `jj diff` would give us
the changes represented by what is being commited, but that isn't true.
In Jujutusu you can (and do!) often update commit messages of changes
that are not the current working copy (making `jj diff` incorrect).

This works well with my normal workflow is to use a custom template for
`jj commit` to include the git diff **or** when using a git repo still
to use `git commit --verbose` which includes the diff.