chore(shells): Flatten shell configuration into packages
Instead of nesting `packages/shells/*`, hoist them up to `packages/`
directly. This means we have to run `cache-shell-startup` multiple times
(because we can't point it at a single path anymore). In theory, we
could update `cache-shell-startup` to accept an array of source<->dest
combos but this seems fine for now.