fix(taskfiles/brew.yml): Ensure `brew bundle` is ran from dotfiles root
Apparently, the default working directory for every task is the
directory of the `Taskfile` that the task is ran from. When `brew.yml`
is included by `Taskfile.dist.yml` in the root, the working directory of
the `brew:update` task is that of the top level `Taskfile.dist.yml` (the
dotfiles repo root), but when `brew.yml` is included in
`taskfiles/system.yml` the working directory is set to `taskfiles/`
which then errors because `brew bundle` requires a `Brewfile` to be in
the working directory.

This fixes that issue by ensuring the working directory is set to the
repo root specifically for `brew:bundle` task.