feat(mise): Install rust and cargo tools via mise
This is limited to things like `cargo-insta`, `cargo-nextest`, and
`cargo-binstall` mostly because for everything we install this way we
don't get completions out of the box. I've already setup completion
integrations in `packages/zsh/` for these ones though.

Note: for `rust` `mise` actually installs rustup for us, which is quite
nice (which means that we still have the same relative pathing for
`rustup` related tools that we did before).