fix(install): Ensure `mise up` is ran as part of `./install`
This is a bit backwards, but it's just really helpful. Since the main
system / platform tools install is always ugprading (because it uses
`brew bundle install`, which updates homebrew database then installs or
upgrades any of the packages), but we didn't have an automated way to
run the `mise up` to get the mise managed tools up to date.

Basically, at this point `./install` is a "one stop shop" to a usable
system.