fix(nvim): Ensure mcphub is lazily loaded
Prior to this change, `mcphub.nvim` would eagerly load when I run
`startup-tmux` which caused `mcp-hub` CLI to run eagerly on all of the
opened nvim instances. Each of the instances would see that there is no
running server on the assigned port then try to start one, which causes
all but one of them to error with `EADDRINUSE`.